rearrange the things
